adding wsdlInformationService11
