+++ lib/ParserFactory.js
@@ -1,3 +1,6 @@
+const { WsdlInformationService11 } = require('./WsdlInformationService11');
+const { WSDLInformationService20 } = require('./WsdlInformationService20');
+
 const V11 = '1.1',
   V20 = '2.0',
   {
@@ -16,9 +19,9 @@ class ParserFactory {
   getParser(xmlDocumentContent) {
     let version = this.getWsdlVersion(xmlDocumentContent);
     if (version === V11) {
-      return new Wsdl11Parser(xmlDocumentContent);
+      return new Wsdl11Parser(new WsdlInformationService11());
     }
-    return new Wsdl20Parser(xmlDocumentContent);
+    return new Wsdl20Parser(new WSDLInformationService20());
   }
 
   /**
